requirements.txt
werkzeug==2.3.7
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.3.0
yfinance==0.2.18
tensorflow==2.13.0
plotly>=5.3.1
dash>=2.0.0
nltk>=3.6.3
textblob>=0.15.3
newsapi-python>=0.2.6
ta>=0.7.0
python-dotenv>=0.19.0
requests>=2.26.0
gunicorn>=20.1.0 